refactor : app layer 리팩토링
