+++ meerket/meerket-application/src/main/java/org/j1p5/MeerketApplication.java
@@ -1,4 +1,4 @@
-package org.j1p5.api;
+package org.j1p5;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -7,8 +7,8 @@ import org.springframework.context.annotation.ComponentScan;
 
 
 @SpringBootApplication
-@ComponentScan(basePackages = "org.j1p5")
-@EntityScan(basePackages = "org.j1p5") // 엔티티 스캔
+//@ComponentScan(basePackages = "org.j1p5")
+//@EntityScan(basePackages = "org.j1p5") // 엔티티 스캔
 public class MeerketApplication {
 
     public static void main(String[] args) {
